+++ src/Workflow/Type/DefaultWorkflowType.php
@@ -17,6 +17,7 @@ use Netzmacht\ContaoWorkflowBundle\PropertyAccess\PropertyAccessManager;
 use Netzmacht\ContaoWorkflowBundle\Workflow\Flow\Action\Integration\UpdateEntityAction;
 use Netzmacht\ContaoWorkflowBundle\Workflow\Flow\Condition\Workflow\PropertyCondition;
 use Netzmacht\Workflow\Flow\Workflow;
+use Netzmacht\Workflow\Manager\Manager;
 use function array_keys;
 
 /**
@@ -34,16 +35,36 @@ final class DefaultWorkflowType extends AbstractWorkflowType
     private $propertyAccessManager;
 
     /**
+     * Workflow manager.
+     *
+     * @var Manager
+     */
+    private $workflowManager;
+
+    /**
+     * Default workflow type configuration.
+     *
+     * @var array
+     */
+    private $defaultWorkflowTypes;
+
+    /**
      * DefaultWorkflowType constructor.
      *
      * @param PropertyAccessManager $propertyAccessManager Property access manager.
+     * @param Manager               $workflowManager       Workflow manager.
      * @param array                 $configuration         Default workflow type configuration.
      */
-    public function __construct(PropertyAccessManager $propertyAccessManager, array $configuration = [])
-    {
+    public function __construct(
+        PropertyAccessManager $propertyAccessManager,
+        Manager $workflowManager,
+        array $configuration = []
+    ) {
         parent::__construct('default_type', array_keys($configuration));
 
         $this->propertyAccessManager = $propertyAccessManager;
+        $this->workflowManager       = $workflowManager;
+        $this->defaultWorkflowTypes  = $configuration;
     }
 
     /**
@@ -57,7 +78,8 @@ final class DefaultWorkflowType extends AbstractWorkflowType
             new PropertyCondition($this->propertyAccessManager, 'workflow', $workflow->getName())
         );
 
-        $action = new UpdateEntityAction($this->propertyAccessManager);
+        $permission = ($this->defaultWorkflowTypes[$workflow->getProviderName()]['step_permission'] ?? false);
+        $action     = new UpdateEntityAction($this->propertyAccessManager, $this->workflowManager, $permission);
         foreach ($workflow->getTransitions() as $transition) {
             $transition->addPostAction($action);
         }